+++ src/CropDetails/CropDetails.css
